refactor(captcha-severity): own the puzzle difficulty ladder
The band table, `severityToPuzzleDifficulty`, `clampDifficultyLevel`,
`MAX_AUTO_ESCALATION_LEVEL` and `MIN_DECOY_HOLE_DARKEN_MARGIN` lived in
`provider/src/tasks/puzzle/puzzleDifficulty.ts`, reachable only from
inside the provider. This package's own docs described the ladder at
length but did not hold it, so anything else that has to answer "what
difficulty is this puzzle" had to restate the mapping — and the
consumers that author and edit puzzle rules sit outside the provider.

Severity already ranks a policy by its `solvedImagesCount`; the ladder
says what a puzzle policy carrying that number is actually served at.
Same question, one step further in.

Sampling a render from a level's bands stays in the provider as
`samplePuzzleDifficulty`: it needs `IPuzzleSettings` and a `node:crypto`
sampler, and this package's zero-dependency, browser-safe property is
load-bearing for consumers that bundle it standalone.

Adds `puzzleDifficultyToSeverity`, the inverse. Writers need that
direction, and each level spans two rounds — without a canonical value
per level a writer picks between numbers that produce the identical
puzzle, and the difference survives only to break severity ties
arbitrarily. The round-trip is pinned by tests.

No behaviour change: same table, same thresholds, same clamping.
