+++ packages/provider/src/tasks/frictionless/frictionlessTasks.ts
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { severityToPuzzleDifficulty } from "@prosopo/captcha-severity";
 import type { Logger } from "@prosopo/logger";
 import { DEFAULT_RENDER_SETTINGS } from "@prosopo/puzzle-assets";
 import {
@@ -48,10 +49,7 @@ import { CaptchaManager } from "../captchaManager.js";
 import { coerceToEnabledCaptchaType } from "../captchaTypeSelection.js";
 import { DecisionMachineRunner } from "../decisionMachine/decisionMachineRunner.js";
 import { getBotScore } from "../detection/getBotScore.js";
-import {
-	samplePuzzleDifficulty,
-	severityToPuzzleDifficulty,
-} from "../puzzle/puzzleDifficulty.js";
+import { samplePuzzleDifficulty } from "../puzzle/puzzleDifficulty.js";
 import { type RoutingContext, applyRouter } from "./routingMachine.js";
 
 const DEFAULT_MAX_TIMESTAMP_AGE = 60 * 10 * 1000; // 10 minutes
